+++ src/imr_intruder/__init__.py
@@ -1,4 +1,4 @@
 """imr-intruder package."""
 
-__version__ = "1.4.1"
+__version__ = "1.4.2"
 APP_NAME = "imr-intruder"